core/requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiogram-calendar==0.5.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.11.0
astroid==3.3.5
asyncpg==0.30.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.1
dill==0.3.9
docker==7.1.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
isort==5.13.2
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.1.0
platformdirs==4.3.6
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pylint==3.3.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
tomlkit==0.13.2
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
yarl==1.17.1